Add Obligation/Duty Tables to Group Manager (#702)
Obligation/Duty tables are only present if a connected player has any
obligations or duties. The table itself is clickable to open up
character sheets same as the player list. The roll buttons are separate
for now even though the only difference is the text displayed so that
additional logic can be added.