Improve FFG dice pool dialog UI
The motivation behind this commit is to improve the readability and reduce the perceived visual complexity of the FFG dice pool dialog.

Centred the dice to give them more room and balance.
Aligned buttons so that they align with the surrounding elements and blocks.
Added spacing between the different sections.
Reordered buttons to upgrade or downgrade abilities and difficulty so that the ability buttons are on the left and the difficulty buttons are on the right. This matches how the dice are displayed and how the section with the added results is ordered.
The text on these buttons is now left-aligned to improve scanning and readability.
The text explaining how to increase or decrease dice is now emphasised and left-aligned.
Left-aligned the text for the various dice for better scanning and readability.
Reordered the results section so that positive (ability) results are on the left and negative (difficulty) are on the right. Reduced this section to two main columns to reduce complexity and improve flow.
Moved audio/chat options to the bottom so that the dice options are grouped together.
Layout no longer breaks when using a language with longer texts.