Feat: Add OggDude's GMTools NPC Importer Tool
This patch introduces an NPC importer tool that is capable of reading and processing exported Adversary XML files from OggDude's GMTools. This is achieved via a new button at the bottom of the Actors tab of FoundryVTT. It has been tested with a number of OggDude's Adversaries as well as my own.

Known Issues:
- Custom Armor, Weapons and Items added in GMTools are currently ignored.
- Adversary's Soak values can be reset if they are granted extra soak in the GMTool that isn't accounted for by standard gear. You can work around this by disabling Soak auto-calculation in the Sheet Options.
- Force powers automatically have every upgrade granted because adversary data does not provide good data for deciding which upgrades have been purchased, only that a force-power is available to that NPC.
- Misc. Abilities that are not Talents are currently written to the Character Biography.
- Talents that allow an NPC to exchange a characteristic used for a skill check are not currently reflected in the Adversary Sheet.