Add adjusted values to template.json
With the new Document object types, only values that are
part of the system template.json are considered valid objects
in the Document#schema. Therefore any elements that are not
part of the schema will not be stored to the database.

This is possibly causing us problems passing adjusted
values around between items, actors, and their various
sheet and roll templates. This might fix it.

Bumped the system version to 1.51 to ensure Foundry
migrates existing items and actors.