Fix item attachments and mods
At least one issue remains, you can't remove base mods (attrs)
from attachments or mods that are embedded on a weapon or armour
item. You can add and remove them just fine on the non-embedded
items though so a suitable workaround for now is to just remove
and re-add the items if you need to change base mods.