Replace usages of window.location.reload with new debounced method
From https://foundryvtt.wiki/en/migrations/foundry-core-0_8_x#reloading-onchange:
```
A change in the way 0.8 processes onChange callbacks in settings as
they are saved has made it inadvisable to use window.location.reload()
directly. Doing so will trigger the refresh and sometimes cause a race
condition that results in some settings not being saved. It is instead
recommended to create a debounced reload callback and use that in each
of your settings.
```