Fix spec talent character sheet rollout.
Sadly this won't fix existing broken specialisations / talents, as the problem was occurring during import.

But it will fix newly imported specialisations and subsequent characters.

The fix for existing characters would be to update the reference between the specialisation and the talent item (by dragging the talent items onto the specialisation).