Handle combatant being undefined
In foundry v9, turn is allowed to be set to null in order to mark that
there is no active combatant. When this is the case, the combatant
getter for the current combatant will return undefined. By option
chaining the name and id properties, this removes an exception that
prevents the dialog from opening.

Additionally, when updating the turn after rolling initiative, prevent
trying to update the turn to -1 and raising an error.

Ref: BoltsJ/lancer-initiative#20, BoltsJ/lancer-initiative#25