Fix anchor references of journal toc
When processing the journal texts and inserting dice icons the whole
HTML of the journal section is replaced, those the element references
of the journal's toc point to elements no longer in the document.
This commit changes the HTML replacement from using outerHTML to
innerHTML as this is somehow necessary for the updated HTMl to be
used by Foundry in subsequent hooks.
This commit also adds another step after the processing of the HTML
in which for each toc item the newly created HTML is searched to find
the data anchors again. The old references are then replaced by the
new ones.

Note that this only fixes a symptom while the root cause is
the replacement of huge chunks the DOM with innerHTML, but the fix
for the symptom is easier to achieve right now.

Fixes #1233