Deprecate the default theme in favour of Mandar theme
We have enough things on life support in this codebase, the CSS styling is one of them. It's grown absurdly messy, confusing and hard to maintain.

Fortunately, it's totally unnecessary for us to maintain that frankenstein's monster when we have a perfectly functional and much neater set of styles in the Mandar theme.

This PR starts the process of deprecating the old styling by first making Mandar the default setting for new worlds.

Follow-up PRs can handle digging out the old mess and/or refactoring it into something new and experimental.